Created logic of two test methods. This proved to be somewhat difficult - I included the Turbine dependency so I would be able to better test flow output sequentially. I wanted to ensure the size of the flow collection, otherwise, I could have easily used the terminal operator .firstOrNull(), which was passing in my test suite prior. I wasn't sure if this was working because the time it took to receive the first item was faster than the assertEquals operations, so it seemed a bit hacky to me to use it. Ultimately, I'm happy with current coroutine set up within the test methods.
