+++ app/build.gradle
@@ -44,6 +44,7 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.5.0'
     implementation 'com.google.android.material:material:1.6.1'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4")
 
     //Navigation
     implementation 'androidx.navigation:navigation-fragment-ktx:2.5.1'
@@ -58,11 +59,14 @@ dependencies {
     implementation 'androidx.test.ext:junit-ktx:1.1.3'
 
     testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.4"
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    implementation("app.cash.turbine:turbine:0.9.0")
 
     implementation("androidx.room:room-runtime:2.4.3")
     annotationProcessor("androidx.room:room-compiler:2.4.3")
     // To use Kotlin annotation processing tool (kapt)
     kapt("androidx.room:room-compiler:2.4.3")
+    implementation "androidx.room:room-ktx:2.4.3"
 }
